# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "博邦 ERP Echo + GORM 模块化单体后台接口文档。后续新增、修改、删除接口时，必须同步更新 Swagger 注释、docs/API.md 和 test.http。",
        "title": "博邦 ERP 管理系统 API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/system/updates/check": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system-updates"
                ],
                "summary": "立即执行完整更新检查",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_update.SystemUpdateStatus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/system/updates/server/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "system-updates"
                ],
                "summary": "下载并校验服务端升级包",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/system/updates/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system-updates"
                ],
                "summary": "查询版本与更新状态",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_update.SystemUpdateStatus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/updates/client/artifacts/{sha256}": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "updates"
                ],
                "summary": "下载已验签客户端更新资源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "当前 manifest 资源 SHA-256",
                        "name": "sha256",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "206": {
                        "description": "Partial Content",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/updates/client/plan": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "updates"
                ],
                "summary": "查询 Windows 客户端完整更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "当前客户端 SemVer",
                        "name": "current_version",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "目标平台，固定 windows-x86_64",
                        "name": "target",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "安装模式：nsis 或 portable",
                        "name": "install_mode",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_update.ClientUpdatePlan"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/updates/client/tauri/{target}/{arch}/{current_version}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "updates"
                ],
                "summary": "查询 Tauri 完整客户端更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tauri target，固定 windows",
                        "name": "target",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CPU 架构，固定 x86_64",
                        "name": "arch",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "当前客户端 SemVer",
                        "name": "current_version",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_update.TauriUpdateResponse"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_update.ClientComponentStatus": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "cached": {
                    "type": "boolean"
                },
                "current_version": {
                    "type": "string"
                },
                "download_path": {
                    "type": "string"
                },
                "download_url": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "latest_version": {
                    "type": "string"
                },
                "manifest": {
                    "$ref": "#/definitions/internal_update.PackageManifest"
                },
                "sha256": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "internal_update.ClientUpdatePlan": {
            "type": "object",
            "properties": {
                "artifact": {
                    "$ref": "#/definitions/internal_update.ClientUpdatePlanArtifact"
                },
                "current_version": {
                    "type": "string"
                },
                "download_size": {
                    "type": "integer"
                },
                "full_size": {
                    "type": "integer"
                },
                "install_mode": {
                    "type": "string"
                },
                "latest_version": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "protocol_version": {
                    "type": "integer"
                },
                "signature": {
                    "type": "string"
                },
                "signed_payload": {
                    "type": "string"
                },
                "strategy": {
                    "type": "string"
                },
                "target": {
                    "type": "string"
                }
            }
        },
        "internal_update.ClientUpdatePlanArtifact": {
            "type": "object",
            "properties": {
                "download_path": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "sha256": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "internal_update.ComponentStatus": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "current_version": {
                    "type": "string"
                },
                "download_path": {
                    "type": "string"
                },
                "download_url": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "latest_version": {
                    "type": "string"
                },
                "manifest": {
                    "$ref": "#/definitions/internal_update.PackageManifest"
                },
                "sha256": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "internal_update.Manifest": {
            "type": "object",
            "properties": {
                "all_in_one": {
                    "$ref": "#/definitions/internal_update.PackageManifest"
                },
                "client_update_v2": {
                    "$ref": "#/definitions/internal_update.SignedClientUpdateManifest"
                },
                "notes": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "server": {
                    "$ref": "#/definitions/internal_update.PackageManifest"
                },
                "updater": {
                    "$ref": "#/definitions/internal_update.PackageManifest"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "internal_update.PackageManifest": {
            "type": "object",
            "properties": {
                "sha256": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "internal_update.SignedClientUpdateManifest": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
        "internal_update.SystemUpdateStatus": {
            "type": "object",
            "properties": {
                "check_interval": {
                    "type": "string"
                },
                "checking": {
                    "type": "boolean"
                },
                "client": {
                    "$ref": "#/definitions/internal_update.ClientComponentStatus"
                },
                "client_cache_bytes": {
                    "type": "integer"
                },
                "client_full_cached": {
                    "type": "boolean"
                },
                "client_protocol_version": {
                    "type": "integer"
                },
                "enabled": {
                    "type": "boolean"
                },
                "interval_seconds": {
                    "type": "integer"
                },
                "last_attempt_at": {
                    "type": "string"
                },
                "last_error": {
                    "type": "string"
                },
                "last_success_at": {
                    "type": "string"
                },
                "manifest": {
                    "$ref": "#/definitions/internal_update.Manifest"
                },
                "manifest_url": {
                    "type": "string"
                },
                "next_check_at": {
                    "type": "string"
                },
                "reachable": {
                    "type": "boolean"
                },
                "server": {
                    "$ref": "#/definitions/internal_update.ComponentStatus"
                }
            }
        },
        "internal_update.TauriUpdateResponse": {
            "type": "object",
            "properties": {
                "pub_date": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT 登录凭证，格式为：Bearer \u003ctoken\u003e。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}